backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
openai
python-dotenv
psycopg2-binary
SQLAlchemy
groq
regex
tenacity
tqdm
langchain
dataclasses
aiohttp
asyncio
typing-extensions
pydantic-settings
markdown
markdown-it-py
mdurl
sseclient-py

# 🔐 AUTH (CRITICAL – DO NOT CHANGE VERSIONS)
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart
pymupdf